refactor: move mixins
