+++ packages/oruga-next/src/mixins/TabbedChildMixin.ts
