+++ packages/oruga-next/src/mixins/TabbedMixin.ts
